refactor(blorktools): glb preview util
